feat: `docker` executable provisioner
Add the provisioner to create a shell script that delegates to docker-run and adds the script to $PATH while in `mod exec`. This gives you `bundle exec`-like experience for any apps in your script that are distribued as docker images.

This feature is backed by k-kinzal/aliases (https://github.com/k-kinzal/aliases). Big kudos to the author!